Added max password retry functionality

+++ android/app/src/main/kotlin/com/example/factory_reset/MainActivity.kt
@@ -1,5 +1,6 @@
 package com.example.factory_reset
 
+import android.app.KeyguardManager
 import android.app.admin.DeviceAdminReceiver
 import android.app.admin.DevicePolicyManager
 import android.content.ComponentName
@@ -19,35 +20,75 @@ class MainActivity : FlutterActivity() {
     private val CHANNEL = "com.example.factory_reset/main"
     private lateinit var mDPM: DevicePolicyManager
     private lateinit var mDeviceAdminComponentName: ComponentName
+
+    private val REQUEST_CODE_ENABLE_ADMIN = 1
+    private val currentAPIVersion = Build.VERSION.SDK_INT
     private val TAG = "Method Channel"
 
     override fun configureFlutterEngine(@NonNull flutterEngine: FlutterEngine) {
         super.configureFlutterEngine(flutterEngine)
-        MethodChannel(
-            flutterEngine.dartExecutor.binaryMessenger,
-            CHANNEL
-        ).setMethodCallHandler { call, result ->
-            Log.d(TAG, "Method Call Name: ${call.method}")
-            if (call.method == "reset") {
-                reset()
-            } else {
-                result.notImplemented()
-            }
 
+        mDPM = getSystemService(Context.DEVICE_POLICY_SERVICE) as DevicePolicyManager
+        mDeviceAdminComponentName = ComponentName(this, DeviceAdmin::class.java)
+
+        MethodChannel(flutterEngine.dartExecutor.binaryMessenger, CHANNEL).setMethodCallHandler {
+                call,
+                result ->
+            Log.d(TAG, "Method Call Name: ${call.method}")
+            try {
+                when (call.method) {
+                    "isDeviceSecured" -> {
+                        val deviceStatus = isDeviceSecured()
+                        result.success(deviceStatus)
+                    }"setMaxPasswordRetries" -> {
+                        val maxRetries = call.argument("maxRetries") as Int?
+                        setMaxFailedPasswordsForWipe(maxRetries)
+                        result.success(true)
+                    }
+                    "getMaxPasswordRetries" -> {
+                        val currentMaxRetries = getMaxFailedPasswordsForWipe()
+                        result.success(currentMaxRetries)
+                    }
+                    "enablePermission" -> {
+                        enablePermission()
+                        result.success(true)
+                    }
+                    "reset" -> {
+                        reset()
+                        result.success(true)
+                    }
+                    else -> {
+                        result.notImplemented()
+                    }
+                }
+            } catch (e: Error) {
+                Log.e(TAG, e.toString())
+                result.error("Platform Error", e.message, null);
+            }
         }
     }
 
     private fun isActiveAdmin(): Boolean = mDPM.isAdminActive(mDeviceAdminComponentName)
-    private val REQUEST_CODE_ENABLE_ADMIN = 1
 
-    private fun reset() {
-        mDPM = getSystemService(Context.DEVICE_POLICY_SERVICE) as DevicePolicyManager
-        mDeviceAdminComponentName = ComponentName(this, DeviceAdmin::class.java)
-//        if (isActiveAdmin()) {
-//            dpm.wipeData(0)
-//        }
-        val currentAPIVersion = Build.VERSION.SDK_INT;
-        Log.d(TAG, "Current API Version: $currentAPIVersion")
+    private fun isDeviceSecured(): Boolean {
+        val keyguardManager = getSystemService(KEYGUARD_SERVICE) as KeyguardManager //api 16+
+        return if (currentAPIVersion >= Build.VERSION_CODES.M) {
+            keyguardManager.isDeviceSecure
+        } else keyguardManager.isKeyguardSecure
+    }
+
+    private fun setMaxFailedPasswordsForWipe(retries: Int?) {
+        if(retries==null || retries<=0) return
+        enablePermission()
+        mDPM.setMaximumFailedPasswordsForWipe(mDeviceAdminComponentName, retries)
+    }
+
+    private fun getMaxFailedPasswordsForWipe(): Int {
+        enablePermission()
+        return mDPM.getMaximumFailedPasswordsForWipe(mDeviceAdminComponentName)
+    }
+
+    private fun enablePermission() {
         if (currentAPIVersion >= Build.VERSION_CODES.FROYO) {
             // 2.2+
             Log.d(TAG, "Is Admin Active: ${isActiveAdmin()}")
@@ -55,17 +96,23 @@ class MainActivity : FlutterActivity() {
                 val intent = Intent(DevicePolicyManager.ACTION_ADD_DEVICE_ADMIN)
                 intent.putExtra(DevicePolicyManager.EXTRA_DEVICE_ADMIN, mDeviceAdminComponentName)
                 intent.putExtra(
-                    DevicePolicyManager.EXTRA_ADD_EXPLANATION,
-                    "Process will remove user installed applications, settings, wallpaper and sound settings. Are you sure you want to wipe device?"
+                        DevicePolicyManager.EXTRA_ADD_EXPLANATION,
+                        "Process will remove user installed applications, settings, wallpaper and sound settings. Are you sure you want to wipe device?"
                 )
                 startActivityForResult(intent, REQUEST_CODE_ENABLE_ADMIN)
+            }
+        }
+    }
+
+    private fun reset() {
+        if (currentAPIVersion >= Build.VERSION_CODES.FROYO) {
+            // 2.2+
+            Log.d(TAG, "Is Admin Active: ${isActiveAdmin()}")
+            if (!isActiveAdmin()) {
+                enablePermission()
             } else {
                 try {
-                    Log.d(TAG, "Wiping Data")
-//                    if (currentAPIVersion <= Build.VERSION_CODES.TIRAMISU){
                     mDPM.wipeData(0)
-//                    }
-                    Log.d(TAG, "Wipe Complete")
                 } catch (e: Error) {
                     Log.e(TAG, e.toString())
                 }
@@ -73,12 +120,13 @@ class MainActivity : FlutterActivity() {
         } else {
             // 2.1
             try {
-                val foreignContext = createPackageContext(
-                    "com.android.settings",
-                    CONTEXT_IGNORE_SECURITY or CONTEXT_INCLUDE_CODE
-                )
+                val foreignContext =
+                        createPackageContext(
+                                "com.android.settings",
+                                CONTEXT_IGNORE_SECURITY or CONTEXT_INCLUDE_CODE
+                        )
                 val yourClass =
-                    foreignContext.classLoader.loadClass("com.android.settings.MasterClear")
+                        foreignContext.classLoader.loadClass("com.android.settings.MasterClear")
                 val i = Intent(foreignContext, yourClass)
                 startActivityForResult(i, REQUEST_CODE_ENABLE_ADMIN)
             } catch (e: ClassNotFoundException) {
@@ -93,10 +141,11 @@ class MainActivity : FlutterActivity() {
         if (requestCode == REQUEST_CODE_ENABLE_ADMIN) {
             if (resultCode == RESULT_OK) {
                 Toast.makeText(
-                    context,
-                    context.getString(R.string.admin_receiver_status_enabled),
-                    Toast.LENGTH_SHORT
-                ).show()
+                                context,
+                                context.getString(R.string.admin_receiver_status_enabled),
+                                Toast.LENGTH_SHORT
+                        )
+                        .show()
             } else {
                 Toast.makeText(context, "Admin Request denied.", Toast.LENGTH_SHORT).show()
             }
@@ -113,12 +162,11 @@ class DeviceAdmin : DeviceAdminReceiver() {
     }
 
     override fun onEnabled(context: Context, intent: Intent) =
-        showToast(context, context.getString(R.string.admin_receiver_status_enabled))
+            showToast(context, context.getString(R.string.admin_receiver_status_enabled))
 
     override fun onDisableRequested(context: Context, intent: Intent): CharSequence =
-        context.getString(R.string.admin_receiver_status_disable_warning)
+            context.getString(R.string.admin_receiver_status_disable_warning)
 
     override fun onDisabled(context: Context, intent: Intent) =
-        showToast(context, context.getString(R.string.admin_receiver_status_disable_warning))
-
+            showToast(context, context.getString(R.string.admin_receiver_status_disable_warning))
 }
